Improve transitions and buttons
