+++ app/src/main/res/layout/app_entry_fragment.xml
@@ -19,6 +19,7 @@
 
 -->
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+             xmlns:app="http://schemas.android.com/apk/res-auto"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
              android:background="@color/teams_background">
@@ -28,13 +29,15 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-    <Button
+    <ImageButton
         android:id="@+id/back_button"
         android:layout_width="@dimen/teams_back_button_size"
         android:layout_height="@dimen/teams_back_button_size"
-        android:layout_margin="@dimen/wire__padding__regular"
         android:layout_gravity="top|start"
-        android:background="@drawable/action_back_dark"/>
-
+        app:srcCompat="@drawable/action_back_dark"
+        android:background="@color/transparent"
+        android:scaleType="fitCenter"
+        android:padding="@dimen/wire__padding__large"
+        />
 
 </FrameLayout>
