+++ app/src/main/res/layout/app_entry_scene.xml
@@ -62,7 +62,8 @@
         android:paddingStart="@dimen/wire__padding__regular"
         android:paddingEnd="@dimen/wire__padding__regular"
         android:background="@drawable/app_entry_button_background"
-        android:clickable="true">
+        android:clickable="true"
+        android:focusable="true">
 
         <com.waz.zclient.ui.text.GlyphTextView
             android:layout_width="@dimen/app_entry_button_size"
@@ -110,7 +111,8 @@
         android:paddingEnd="@dimen/wire__padding__regular"
         android:layout_marginTop="@dimen/wire__padding__regular"
         android:background="@drawable/app_entry_button_background"
-        android:clickable="true">
+        android:clickable="true"
+        android:focusable="true">
 
         <com.waz.zclient.ui.text.GlyphTextView
             android:layout_width="@dimen/app_entry_button_size"
@@ -151,26 +153,40 @@
         android:layout_height="0dp"
         android:layout_weight="0.8"/>
 
-    <com.waz.zclient.ui.text.TypefaceTextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:textSize="@dimen/wire__text_size__regular_small"
-        android:textColor="@color/teams_description_text"
-        app:w_font="@string/wire__typeface__regular"
-        android:text="@string/app_entry_login_sub"/>
-
-    <com.waz.zclient.ui.text.TypefaceTextView
+    <LinearLayout
         android:id="@+id/login_button"
-        android:layout_width="wrap_content"
+        android:layout_width="@dimen/app_entry_button_width"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:textSize="@dimen/wire__text_size__medium"
-        android:textColor="@color/teams_headline_text"
-        app:w_font="@string/wire__typeface__bold"
-        android:text="@string/app_entry_login"/>
+        android:orientation="vertical"
+        android:layout_gravity="center_horizontal"
+        android:paddingTop="@dimen/wire__padding__regular"
+        android:paddingBottom="@dimen/wire__padding__regular"
+        android:background="@drawable/app_entry_button_background"
+        android:clickable="true"
+        android:focusable="true">
+
+        <com.waz.zclient.ui.text.TypefaceTextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:textSize="@dimen/wire__text_size__regular_small"
+            android:textColor="@color/teams_description_text"
+            app:w_font="@string/wire__typeface__regular"
+            android:text="@string/app_entry_login_sub"/>
+
+        <com.waz.zclient.ui.text.TypefaceTextView
+
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:gravity="center"
+            android:textSize="@dimen/wire__text_size__medium"
+            android:textColor="@color/teams_headline_text"
+            app:w_font="@string/wire__typeface__bold"
+            android:text="@string/app_entry_login"/>
+
+    </LinearLayout>
 
     <View
         android:layout_width="match_parent"
