+++ app/src/main/scala/com/waz/zclient/appentry/scenes/FirstScreenViewHolder.scala
@@ -18,29 +18,22 @@
 package com.waz.zclient.appentry.scenes
 
 import android.content.Context
-import android.view.View.{OnClickListener, OnTouchListener}
 import android.view._
 import android.widget.LinearLayout
-import com.waz.ZLog
 import com.waz.utils.events.EventContext
 import com.waz.zclient._
+import com.waz.zclient.appentry.AppEntryButtonOnTouchListener
 import com.waz.zclient.controllers.SignInController
 import com.waz.zclient.controllers.SignInController._
-import com.waz.zclient.ui.text.TypefaceTextView
-import android.animation.Animator
-import android.graphics.Rect
-import com.waz.zclient.appentry.scenes.FirstScreenViewHolder.AppEntryButtonOnTouchListener
 
 case class FirstScreenViewHolder(root: View)(implicit val context: Context, eventContext: EventContext, injector: Injector) extends ViewHolder with Injectable {
 
   val appEntryController = inject[AppEntryController]
   val signInController = inject[SignInController]
 
-  //val root = LayoutInflater.from(context).inflate(R.layout.app_entry_scene, container, false).asInstanceOf[ViewGroup]
-
   lazy val createTeamButton = root.findViewById[LinearLayout](R.id.create_team_button)
   lazy val createAccountButton = root.findViewById[LinearLayout](R.id.create_account_button)
-  lazy val loginButton = root.findViewById[TypefaceTextView](R.id.login_button)
+  lazy val loginButton = root.findViewById[LinearLayout](R.id.login_button)
 
   def onCreate(): Unit = {
     createAccountButton.setOnTouchListener(AppEntryButtonOnTouchListener({
@@ -49,61 +42,10 @@ case class FirstScreenViewHolder(root: View)(implicit val context: Context, even
         signInController.uiSignInState ! SignInMethod(Register, Phone)
     }))
     createTeamButton.setOnTouchListener(AppEntryButtonOnTouchListener(() => appEntryController.createTeam()))
-
-    loginButton.setOnClickListener(new OnClickListener {
-        override def onClick(v: View): Unit = {
-          appEntryController.goToLoginScreen()
-          signInController.uiSignInState ! SignInMethod(Login, Email)
-        }
-      })
-  }
-}
-
-object FirstScreenViewHolder {
-  case class AppEntryButtonOnTouchListener(onClick: () => Unit)(implicit context: Context) extends OnTouchListener {
-
-    private var finished = false
-
-    def isOutside(v: View, event: MotionEvent): Boolean ={
-      val rect = new Rect(v.getLeft, v.getTop, v.getRight, v.getBottom)
-      !rect.contains(v.getLeft + event.getX.toInt, v.getTop + event.getY.toInt)
-    }
-
-    def animatePress(v: View): ViewPropertyAnimator =
-      v.animate()
-        .scaleX(0.96f)
-        .scaleY(0.96f)
-        .setDuration(context.getResources.getInteger(android.R.integer.config_shortAnimTime))
-
-    def animateRelease(v: View): ViewPropertyAnimator =
-      v.animate()
-        .scaleX(1f)
-        .scaleY(1f)
-        .setDuration(context.getResources.getInteger(android.R.integer.config_shortAnimTime))
-
-
-    override def onTouch(v: View, event: MotionEvent): Boolean = {
-      event.getAction match {
-        case MotionEvent.ACTION_DOWN =>
-          animatePress(v)
-          finished = false
-        case MotionEvent.ACTION_MOVE if isOutside(v, event) && !finished =>
-          animateRelease(v).start()
-          finished = true
-        case MotionEvent.ACTION_OUTSIDE | MotionEvent.ACTION_HOVER_EXIT | MotionEvent.ACTION_CANCEL if !finished =>
-          animateRelease(v).start()
-          finished = true
-        case MotionEvent.ACTION_UP if !finished =>
-          finished = true
-          animateRelease(v).setListener(new Animator.AnimatorListener {
-            override def onAnimationEnd(animation: Animator): Unit = onClick()
-            override def onAnimationCancel(animation: Animator): Unit = {}
-            override def onAnimationRepeat(animation: Animator): Unit = {}
-            override def onAnimationStart(animation: Animator): Unit = {}
-          }).start()
-        case _ =>
-      }
-      false
-    }
+    loginButton.setOnTouchListener(AppEntryButtonOnTouchListener({
+      () =>
+        appEntryController.goToLoginScreen()
+        signInController.uiSignInState ! SignInMethod(Login, Email)
+    }))
   }
 }
